pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "groq>=0.22.0",
    "ipykernel>=6.29.5",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
]
